v1.4.0: Major refactor
Instead of sending every incoming message/event to their own
handler (and therefore their own goroutine) we now have a single
"command handler", which differentiates the incoming message in
a switch statement. This helps reducing the memory footprint and
the amount of GCs

Also the code was straightened a lot, so that duplicate code
fragments have been removed.