v1.4.4-RC2: Major overhaul of the RAT cookie handling
The check and notify user if RAT cookie failed logic has been removed from the updateBalance()
method. Instead there is a hourly job (with a random up to 10 minute delay) that will check the
validity of the RAT cookie. If it fails 4 times in a row with a HTTP 403, it is likely that the
cookie is expired or invalid. In this case, a notification DM is sent to the user.