v1.4.4-RC3: User preference encryption added
So far, the SoT RAT cookie was stored as plain text in the database. Starting with v1.4.4-RC3, the
bot has general encryption capabilities. The encryption key is generated at first start up and is
stored in `$HOME/.sotbot_enc_key`. If the file is already present, the content of this file will be
read and used as AES256 encryption key. So far this feature is only used for storing and reading
the SoT RAT cookie.