v1.4.5-RC2: Add trading routes caching (issue #34)
Since the trading routes API seems to be not stable, a local cache for the TRs has been
introduced. To be future-proof a "bot cache" has been added. This is a general cache that
the bot can use to easily serialize objects and store them as base64 encoded string in the
DB for later use. Since the playsot_handler.go makes use of serialization as well, it's
code has been adjusted to make use of the new cache methods as well. To decide if the
TRs are still valid, the TR dates are now parsed and stored as "ValidFrom" and "ValidThru"
fields in the TR struct