#48 smarter summary collection
If a user had no valid RAT cookie during the last summary collection run, the summary data of
that user would not be updated for 24h. To avoid this, the summary collection is now user based
and therefore once a potentially invalid cookie was updated the summary data is updated at latest
30 minutes later in the next run. Also there will be an error log entry if the cookie of a user
is expired.