v1.4.7-RC4: Regularly check if the RAT cookie is still valid
The bot will check every 30 mins. if the registered RAT cookie of a specific user is still valid.
Is the cookie expired, the bot will notify the user one via DM and set a state-flag in the database,
so it will not create multiple DMs every 30 mins. Is a new RAT cookie registered, the flag will be
deleted.