v1.4.8-DEV.1: Force summary collection
The summary collection is running every 24h. This can cause issues for users in different time
zones. To mitigate this, a `/collectsummary` command has been added, so that a user is able to
force a collection of the current summary data that will be used to compare with the next time
`/summary` is executed.